Initial commit of redesigned internal processing and audio management.
Note: Several of the decoders are currently non-functioning in this
commit.  If you want the latest functioning codebase, please checkout
the commit prior to this one.  All decoders will eventually be updated
with this new architecture.

Currently broken:
- Logging and recording
- Aux decoders (Fleetsync/MDC/LoJack/Tait)
- LTR/Passport decoders

Changes in this commit:
- Module architecture for decoders and components with central
provisioning by the DecoderFactory
- Transition from Complex sample stream processing to float buffer
processing for improved memory management
- Redesigned filter structures - to support buffers and streams
- Resolved several memory leak issues
- Universal sharing of the thread pool manager
- Added centralized audio management with call/user prioritization and
new gui with mute/volume controls
- Embedded audio metadata for eventual streaming support
- Redesigned ProcessingChain to support independent re-use
- Added new TrafficChannelManager to handle traffic channel allocations
- Traffic Channel pool size and Call timeout available in P25 and
MPT1327 configs
- Redesigned channel and decoder gui components

+++ src/dsp/filter/FloatFIRFilter.java
@@ -27,13 +27,13 @@ public class FloatFIRFilter implements RealSampleListener
 	private ArrayList<Float> mBuffer;
     private int mBufferSize = 1; //Temporary initial value
 	private int mBufferPointer = 0;
-	private double[] mCoefficients;
+	private float[] mCoefficients;
 	private int[][] mIndexMap;
 	private int mCenterCoefficient;
 	private int mCenterCoefficientMapIndex;
-	private double mGain;
+	private float mGain;
 
-	public FloatFIRFilter( double[] coefficients, double gain )
+	public FloatFIRFilter( float[] coefficients, float gain )
 	{
 		mCoefficients = coefficients;
 		mBuffer = new ArrayList<Float>();
@@ -86,7 +86,7 @@ public class FloatFIRFilter implements RealSampleListener
 
 		//Convolution - multiply filter coefficients by the circular buffer 
 		//samples to calculate a new filtered value
-		double accumulator = 0.0d;
+		float accumulator = 0.0f;
 
 		//Start with the center tap value
 		accumulator += mCoefficients[ mCenterCoefficient ] * 
@@ -103,7 +103,7 @@ public class FloatFIRFilter implements RealSampleListener
 
 		//We're almost finished ... apply gain, cast the doubles to shorts and
 		//return the value
-		return (float)( accumulator * mGain );
+		return accumulator * mGain;
 	}
 
 	/**
